+++ __tests__/browser/feedback/vs.glsl
@@ -3,8 +3,10 @@
 in vec4 vecA;
 in vec4 vecB;
 
-out vec4 result;
+out vec4 result1;
+out vec4 result2;
 
 void main() {
-  result = vecA + vecB;
+  result1 = vecA + vecB;
+  result2 = vecA - vecB;
 }