+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.0
+
+* \[Breaking Change\] Buffers treat dataOrLength as length rather than byte length when its value is number.
+
+* \[Breaking Change\] TransformFeedback.feedback() receives an array of targetBuffers rather than single buffer.
+
 ## 0.0.1
 
 * Initial release with basic functions.